The user interface is basic.
The current UTC time is displayed in the title.

"TeslaSenti" is displayed at the top left.

A graph of sentiment index is displayed, which has yaxis limits from -5 to +5.
The model outputs sentiment index which is -1, 0 or 1.  The number of friends, followers,
retweets etc can be used to calculate a weight and multiply with -1 or 1 as the case might be.
This is yet to be done. The xaxis shows date and time which is 5 minutes apart. This is controlled
by the dtick parameter (see app.py) which has units of microseconds.

Below the graph the latest four tweets are displayed. This page is refreshed every 2 minutes.
This is controlled by the "content" parameter which has units of seconds.