The cleaned tweet (tidy_tweet) is read in along with id, date, tweet,senti and screen_name. The cleaning is done in sentiment_tweets().  As such tidy_tweet is not required in app.py. But is is printed out along with other features to make sure that senti_tweets.csv is correctly read in.
The date column is converted into list of datetime objects which is used in give_graph() as the xaxis. This graph is sent to the front end as a JSON object along with four latest tweets.

Note that app.py uses multiprocessing. The streaming tweets are collected by live_stream().
These are cleaned and model is run by sentiment_tweets(). The requests from the front end
are handled by home(). These three processes run in parallel. These three tasks should
not be interrupted and hence multiprocessing is used. Also these are inherently parallel except for the requirement that streaming_tweets.csv should exist for the cleaning
of the tweets to happen in sentiment_tweets() and  senti_tweets.csv should exist for home() to handle requests. A simple sleep() mechanism is used to make sure that the csv files exist before the code proceeds.