live_stream() collects the streaming tweets and writes to csv file. sentiment_tweets() cleans the tweets, runs the model and writes these to another csv file.
skiprows keeps track of rows already read in from streaming_tweets_save.csv. After every 20 seconds the latest tweets are read in from this file and processed only if there is atleast one new tweet.

The cleaning consists of removing emojis and urls, removing hashtags, cashtags and mentions. Lastly the tokens in the tweets are stemmed.