Test code for displaying two graphs side by side.
The left side graph shows the longer timescale (time_horizon1) while the right side graph shows the shorter timescale (time_horizon2). Both graph show the sentiment index on the yaxis.

senti_tweets.csv is read in. The subset of tweets within time_horizon1 is filtered in. Fro example,  if time_horizon1 is 600 minutes then all tweets within past 10 hours are included. The dates and sentiment indices of these are written to dates1 and arr1. This dataframe is further subseted to include only tweets within time_horizon2 whose dates and sentiment indices are are written to dates2 and arr2. These are passed to give_graph() which returns a figure with two subplots.