The sentiment index is averaged over time intervals (of 1 hour and 5 minutes). The series of these averages are shown along with the individual sentiment indices. This summary over a time interval is more informative since it condenses all the tweets over this time interval into one number.
The dataframe of tweets and sentiment indices is passed to give_graph() rather than passing the individual series of datetime objects and sentiment indices. The averaging is done for the short and long timeseries. The solid line shown the averages. The resample function returns NaN for intervals over which no tweets exist. The graph shows breaks in the solid line over such intervals.