No need for two dataframes dftw and dftw_senti. Since many columns (# retweets, # freinds, etc) are required for weighting dftw is used to read in the latest tweets and tidy_tweet, senti, wt_senti columns are filled in.
Use dftw.empty() rather than checking dftw.shape[0], which is more readable.

Make sure dftw in non-empty before running the sentiment model on dftw. It is
possible that after tidy_tweets with zero tokens are dropped dftw is empty.

skiprows are used to read in the latest tweets which are processed and their sentiment is written to senti_tweets.csv. This can be made more elegant with sql database.