change correlated word
chi2  test is used to find out if two variables are related. The relation can be that of correlation or anticorrelation. Both correlation and anticorrelation contain information. The null hypothesis of chi2 test is that the variables are independent. If the null hypothesis is rejected then it means that the variables are dependent. Hence it is more accurate to say the "top few related words" rather than "top few correlated words".