Requests is an easy-to-use HTTP library for python. requests.get() returns response object. The built in JSON decoder json() is used to flesh out the html content. get() and json() can fail separately and hence it is better to call them separately.
Therefore the first try only catches exceptions thrown by get(). The json is decoded only if get() is successful. get() can throw several different types of exceptions like HTTPError, ConnectionError, ProxyError, SSLError, Timeout error, TooManyRedirects, etc . It is important to  handle them separately in production. For example with ConnectionError, one could retry a few times before bailing out.

In case get() or json(), "tweet not found" is displayed on the frontend which is not desirable. In a more sophisticated approach either the tweet can be dropped or depending on the exception one could try to recover the tweet. The details about the tweet like screenname and id are printed to terminal. In production these could be logged to a file along with the type of exception for offline/retrospective analysis of the missing tweet.

The displayed graph is trimmed down to only one curve which shows the hourly/5-minute average sentiment index.